--- example/lib/src/animated_icon_items.dart
@@ -1,28 +1,0 @@
-import 'package:yaru_icons/yaru_icons.dart';
-
-import 'common/icon_item.dart';
-
-final List<IconItem> animatedIconItems = [
-  IconItem(
-    name: 'Ok',
-    usage: 'YaruAnimatedOkIcon()',
-    iconBuilder: (context, iconSize) => YaruAnimatedOkIcon(
-      size: iconSize,
-    ),
-  ),
-  IconItem(
-    name: 'Ok filled',
-    usage: 'YaruAnimatedOkIcon(filled: true)',
-    iconBuilder: (context, iconSize) => YaruAnimatedOkIcon(
-      size: iconSize,
-      filled: true,
-    ),
-  ),
-  IconItem(
-    name: 'No network',
-    usage: 'YaruAnimatedNoNetworkIcon()',
-    iconBuilder: (context, iconSize) => YaruAnimatedNoNetworkIcon(
-      size: iconSize,
-    ),
-  ),
-];
