+++ example/lib/src/provider/search_provider.dart
@@ -1,6 +1,5 @@
 import 'package:flutter/material.dart';
 
-import '../animated_icon_items.dart';
 import '../common/icon_item.dart';
 import '../icon_items.dart';
